update: code refactoring
